Customerchurn/requirements.txt
pandas
numpy
scikit-learn==1.6.1
matplotlib
seaborn
pyyaml
Flask
gunicorn
scipy
mlflow
dagshub
pytest